Rename `comment-delimiter-utils.js` to `comment-utils.js`
